# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SolarSentinel API — solar flare nowcasting & forecasting
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FluxDataPoint {
  timestamp: string;
  solexs_flux: number;
  hel1os_flux: number;
  /** @nullable */
  solexs_deriv?: number | null;
  /** @nullable */
  hel1os_deriv?: number | null;
  /** @nullable */
  flux_ratio?: number | null;
  /** @nullable */
  spectral_hardness?: number | null;
  /** @nullable */
  rolling_mean?: number | null;
  /** @nullable */
  rolling_var?: number | null;
}

export interface StreamResponse {
  data: FluxDataPoint[];
  total: number;
  current_index: number;
}

export interface DataRangeResponse {
  data: FluxDataPoint[];
}

export interface FeatureContributions {
  /** @nullable */
  flux_derivative?: number | null;
  /** @nullable */
  flux_ratio?: number | null;
  /** @nullable */
  spectral_hardness?: number | null;
  /** @nullable */
  rolling_variance?: number | null;
}

export interface FlareEvent {
  id: string;
  start_time: string;
  peak_time: string;
  end_time: string;
  peak_solexs_flux: number;
  peak_hel1os_flux: number;
  goes_class: string;
  duration_minutes: number;
  detection_confidence: number;
  detection_method: string;
  feature_contributions?: FeatureContributions;
  /** @nullable */
  forecast_probability_at_onset?: number | null;
}

export interface FlareEventsResponse {
  events: FlareEvent[];
}

export interface ForecastOutput {
  timestamp: string;
  probability_15min: number;
  predicted_class: string;
  model: string;
  /** @nullable */
  confidence_interval_low?: number | null;
  /** @nullable */
  confidence_interval_high?: number | null;
}

export interface FeatureSnapshot {
  timestamp: string;
  solexs_flux: number;
  hel1os_flux: number;
  /** @nullable */
  flux_ratio?: number | null;
  /** @nullable */
  spectral_hardness?: number | null;
  /** @nullable */
  solexs_deriv?: number | null;
  /** @nullable */
  rolling_mean?: number | null;
  /** @nullable */
  rolling_var?: number | null;
}

export interface ReplayStatus {
  status: string;
}

export interface ModelMetrics {
  precision: number;
  recall: number;
  f1_score: number;
  tss: number;
  pod: number;
  far: number;
  test_period: string;
  total_flares_tested: number;
  model_version: string;
}

export interface UploadResult {
  status: string;
  rows: number;
  events_detected: number;
}

export type GetDataRangeParams = {
start?: number;
end?: number;
};