# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SolarSentinel API — solar flare nowcasting & forecasting
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /data/range:
    get:
      operationId: getDataRange
      tags: [data]
      summary: Get data for a time index range
      parameters:
        - name: start
          in: query
          schema: { type: integer, default: 0 }
        - name: end
          in: query
          schema: { type: integer, default: 60 }
      responses:
        "200":
          description: Data range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataRangeResponse"
...
  /upload:
    post:
      operationId: uploadData
      tags: [data]
      summary: Upload CSV data for analysis
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Bad request
components:
  schemas:
...
    FluxDataPoint:
      type: object
      required: [timestamp, solexs_flux, hel1os_flux]
      properties:
        timestamp:
          type: string
        solexs_flux:
          type: number
        hel1os_flux:
          type: number
        solexs_deriv:
          type: ["number", "null"]
        hel1os_deriv:
          type: ["number", "null"]
        flux_ratio:
          type: ["number", "null"]
        spectral_hardness:
          type: ["number", "null"]
        rolling_mean:
          type: ["number", "null"]
        rolling_var:
          type: ["number", "null"]
...
    DataRangeResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FluxDataPoint"
...
    UploadResult:
      type: object
      required: [status, rows, events_detected]
      properties:
        status:
          type: string
        rows:
          type: integer
        events_detected:
          type: integer