# lib/api-zod/src/generated/types/dataRangeResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SolarSentinel API — solar flare nowcasting & forecasting
 * OpenAPI spec version: 0.1.0
 */
import type { FluxDataPoint } from './fluxDataPoint';

export interface DataRangeResponse {
  data: FluxDataPoint[];
}