# lib/api-zod/src/generated/types/fluxDataPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SolarSentinel API — solar flare nowcasting & forecasting
 * OpenAPI spec version: 0.1.0
 */

export interface FluxDataPoint {
  timestamp: string;
  solexs_flux: number;
  hel1os_flux: number;
  /** @nullable */
  solexs_deriv?: number | null;
  /** @nullable */
  hel1os_deriv?: number | null;
  /** @nullable */
  flux_ratio?: number | null;
  /** @nullable */
  spectral_hardness?: number | null;
  /** @nullable */
  rolling_mean?: number | null;
  /** @nullable */
  rolling_var?: number | null;
}