# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SolarSentinel API — solar flare nowcasting & forecasting
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  status: string;
  rows: number;
  events_detected: number;
}